+++ modules/Core/include/Open3D/Core/Geometry/PointCloud.h
@@ -100,7 +100,7 @@ std::shared_ptr<PointCloud> CreatePointCloudFromDepthImage(
         const Image &depth, const PinholeCameraIntrinsic &intrinsic,
         const Eigen::Matrix4d &extrinsic = Eigen::Matrix4d::Identity(),
         double depth_scale = 1000.0, double depth_trunc = 1000.0,
-        int stride = 1);
+        int32_t stride = 1);
 
 /// Factory function to create a pointcloud from an RGB-D image and a camera
 /// model (PointCloudFactory.cpp)
@@ -125,7 +125,7 @@ std::shared_ptr<PointCloud> VoxelDownSample(const PointCloud &input,
 /// Function to downsample \param input pointcloud into output pointcloud uniformly
 /// \param every_k_points indicates the sample rate.
 std::shared_ptr<PointCloud> UniformDownSample(const PointCloud &input,
-        size_t every_k_points);
+        uint32_t every_k_points);
 
 /// Function to crop \param input pointcloud into output pointcloud
 /// All points with coordinates less than \param min_bound or larger than
